sdk: Refactor AbstractObjectProvider
Previously, the `AbstractObjectProvider` only worked with
`Identifiables`, which made it incompatible with the new version of the
AAS metamodel.

This makes both the `AbstractObjectProvider` and the
`AbstractObjectStore` independent of `Identifiable`. Furthermore, in all
non-abstract subclasses where `Object` appears in the class or method
name, it has been replaced with `Identifiable`. Old classes and methods
are still available, but with a deprecation warning.

Moreover, a bug in the `LocalFileIdentifiableStore` was identified and
resolved with a quick fix. It will be fixed properly in a separate PR.

Overall, this PR is intended to serve as a basis for discussion, as the
current solution is neither ideal nor final. Therefore, the
documentation has not yet been adapted.

Fixes #428